Move setup into subfolder
Otherwise mocha thinks it’s an actual test
