+++ test/toolbar.js
@@ -1,6 +1,6 @@
 /* eslint-env mocha */
 
-import './setup';
+import './util/setup';
 import {expect} from 'chai';
 import $renderApplication from './util/renderApplication';
 import $ from 'jquery';
