+++ test/util/setup.js
@@ -9,14 +9,14 @@ import 'chai-react';
 import proxyquire from 'proxyquire';
 import {DOMParser} from 'xmlshim';
 import wrap from 'lodash/wrap';
-import mockFirebase from './util/mockFirebase';
-import initI18n from '../src/util/initI18n';
+import mockFirebase from '../util/mockFirebase';
+import initI18n from '../../src/util/initI18n';
 
 chai.use(chaiAsPromised);
 chai.use(dirtyChai);
 chai.use(chaiThings);
 
-proxyquire('../src/services/appFirebase', {
+proxyquire('../../src/services/appFirebase', {
   firebase() {
     return mockFirebase;
   },
